pyproject.toml
[project]
name = "beacon"
version = "0.1.0"
description = "Modular FastAPI alert service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "PTH"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
